[main] clean-up pixel conversion.
The previous on-the-fly pixel conversion was a temporary hack with
some issues. For instance, modifying a const buffer provided by
libretro. I refactored this code path to be cleaner, hack free; at
the expense of an extra copy. This is fine though since it's just
a prototype running on a powerful desktop.